website/pyproject.toml
[tool.poetry]
name = "myTenderWebsite"
version = "0.1.0"
description = ""
authors = ["Alex"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-split-settings = "^1.3.0"
pyyaml = "^6.0.1"
django-storages = "^1.14.2"
boto3 = "^1.34.65"
daphne = "^4.1.0"
channels = "^4.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
colorlog = "^6.8.2"

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
